Add support for * in subject access review requests (#347)
Fields Resource, Group and Verb can be * in SAR requests. Currently there is no handling for such requests.
We will now create a map containing list of predefined K8s types on startup and use that map to create list of data actions that need to be sent in a checkaccess request depending on the combo of *'s in Resource, Group and Verb fields.